fix ambulance type

+++ src/main/java/com/ciphertext/opencarebackend/dto/request/AmbulanceRequest.java
@@ -6,20 +6,11 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AmbulanceRequest {
-
-
-//    private Long id;
     private String vehicleNumber;
     private String type;
     private String driverName;
     private String driverPhone;
-//    private Boolean isAvailable;
-    private Long hospitalId;
-    private Long upazilaId;
-    private Long districtId;
-//    private Boolean isActive;
-//    private String createdBy;
-//    private LocalDateTime createdAt;
-//    private String updatedBy;
-//    private LocalDateTime updatedAt;
+    private Integer hospitalId;
+    private Integer upazilaId;
+    private Integer districtId;
 }
